+++ product_code/certificate_service/api/src/main/scala/de/upb/cs/uc4/certificate/model/EncryptedPrivateKey.scala
@@ -1,6 +1,6 @@
 package de.upb.cs.uc4.certificate.model
 
-import de.upb.cs.uc4.shared.client.configuration.RegexCollection
+import de.upb.cs.uc4.shared.client.configuration.{ ErrorMessageCollection, RegexCollection }
 import de.upb.cs.uc4.shared.client.exceptions.SimpleError
 import play.api.libs.json.{ Format, Json }
 
@@ -13,6 +13,10 @@ case class EncryptedPrivateKey(key: String, iv: String, salt: String) {
     val ivRegex = RegexCollection.EncryptedPrivateKey.ivRegex
     val saltRegex = RegexCollection.EncryptedPrivateKey.saltRegex
 
+    val keyMessage = ErrorMessageCollection.EncryptedPrivateKey.keyMessage
+    val ivMessage = ErrorMessageCollection.EncryptedPrivateKey.ivMessage
+    val saltMessage = ErrorMessageCollection.EncryptedPrivateKey.saltMessage
+
     if (key.isEmpty && iv.isEmpty && salt.isEmpty) {
       Seq()
     }
@@ -23,13 +27,13 @@ case class EncryptedPrivateKey(key: String, iv: String, salt: String) {
       var errors = List[SimpleError]()
 
       if (!keyRegex.matches(key)) {
-        errors :+= SimpleError("key", "For a non-empty key object, the key must not be longer than 8192 characters.")
+        errors :+= SimpleError("key", keyMessage)
       }
       if (!ivRegex.matches(iv)) {
-        errors :+= SimpleError("iv", "For a non-empty key object, the iv must not be longer than 64 characters.")
+        errors :+= SimpleError("iv", ivMessage)
       }
       if (!saltRegex.matches(salt)) {
-        errors :+= SimpleError("salt", "For a non-empty key object, the salt must not be longer than 256 characters.")
+        errors :+= SimpleError("salt", saltMessage)
       }
 
       errors