+++ product_code/configuration_service/impl/src/main/scala/de/upb/cs/uc4/configuration/impl/ConfigurationServiceImpl.scala
@@ -17,11 +17,11 @@ import play.api.Environment
 import scala.concurrent.duration._
 import scala.concurrent.{ ExecutionContext, Future }
 
-/** Implementation of the CourseService */
+/** Implementation of the ConfigurationService */
 class ConfigurationServiceImpl(override val environment: Environment)(implicit ec: ExecutionContext, config: Config)
   extends ConfigurationService {
 
-  implicit val timeout: Timeout = Timeout(15.seconds)
+  implicit val timeout: Timeout = Timeout(config.getInt("uc4.timeouts.hyperledger").milliseconds)
 
   /** Get hyperledger versions */
   override def getHyperledgerVersions: ServiceCall[NotUsed, HyperledgerVersions] = ServiceCall { _ =>