+++ product_code/examreg_service/impl/src/main/scala/de/upb/cs/uc4/examreg/impl/commands/ExamregCommand.scala
@@ -1,0 +1,7 @@
+package de.upb.cs.uc4.examreg.impl.commands
+
+/** The trait for the commands needed in the state
+  * Every command is a case class containing the
+  * necessary information to execute the command
+  */
+trait ExamregCommand extends ExamregCommandSerializable