+++ product_code/authentication_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [v0.13.0](https://github.com/upb-uc4/University-Credits-4.0/compare/authentication-v0.12.0...authentication-v0.13.0) (2020-11-23)
+## Feature
+ - Added configurable timeouts in application config
+## Refactor
+## Bugfix
+
 # [v0.12.0](https://github.com/upb-uc4/University-Credits-4.0/compare/authentication-v0.11.0...authentication-v0.12.0) (2020-11-10)
 ## Feature
  - Added support for ETags