+++ product_code/course_service/CHANGELOG.md
@@ -1,3 +1,19 @@
+# [v0.12.2](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.12.1...course-v0.12.2) (2020-11-17)
+## Feature
+## Refactor
+## Bugfix
+ - Fixed query parameter in getAllCourses
+ - Fixed addCourse fetching all modules instead of only relevant ones
+ 
+# [v0.12.1](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.12.0...course-v0.12.1) (2020-11-17)
+## Feature
+ - Added configurable timeouts in application config
+ - Added "moduleIds" to Course object
+ - Added query parameter for module Ids
+ - Added dependency; CourseService dependant on ExamReg
+## Refactor
+## Bugfix
+
 # [v0.12.0](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.11.1...course-v0.12.0) (2020-11-10)
 ## Feature
 ## Refactor