+++ product_code/user_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [v0.13.0](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.12.0...user-v0.13.0) (2020-11-23)
+## Feature
+ - Added configurable timeouts in application config
+## Refactor
+## Bugfix
+
 # [v0.12.0](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.11.1...user-v0.12.0) (2020-11-10)
 ## Feature
 ## Refactor